refactor portfolio performance route
