+++ src/app/router_registry.py
@@ -28,6 +28,7 @@ from app.routers.foundation import router as foundation_router
 from app.routers.intake import router as intake_router
 from app.routers.platform import router as platform_router
 from app.routers.portfolio import router as portfolio_router
+from app.routers.portfolio_performance import router as portfolio_performance_router
 from app.routers.proposal_delivery import router as proposal_delivery_router
 from app.routers.proposal_memos import router as proposal_memos_router
 from app.routers.proposal_narratives import router as proposal_narratives_router
@@ -65,6 +66,7 @@ def register_routers(app: FastAPI) -> None:
     app.include_router(intake_router)
     app.include_router(foundation_router)
     app.include_router(portfolio_router)
+    app.include_router(portfolio_performance_router)
     app.include_router(composite_performance_router)
     app.include_router(dpm_command_center_router)
     app.include_router(dpm_command_center_outcome_reviews_router)
