+++ tests/integration/test_portfolio_router.py
@@ -2,17 +2,19 @@ import pytest
 from fastapi.testclient import TestClient
 
 from app.main import app
-from app.routers import portfolio as portfolio_router
-from app.routers.portfolio import _portfolio_service
+from app.services.portfolio_service_provider import (
+    portfolio_performance_workspace_service,
+    portfolio_service,
+)
 
 LOTUS_CORE_QUERY_CLIENT = "app.clients.lotus_core_query_client.LotusCoreQueryClient"
 
 
 @pytest.fixture(autouse=True)
 def clear_portfolio_router_cache():
-    _portfolio_service().clear_upstream_cache()
+    portfolio_service().clear_upstream_cache()
     yield
-    _portfolio_service().clear_upstream_cache()
+    portfolio_service().clear_upstream_cache()
 
 
 @pytest.fixture(autouse=True)
@@ -926,7 +928,7 @@ def test_portfolio_insights_router(monkeypatch):
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.get_portfolio_positions", _positions)
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.query_asset_allocation", _allocation)
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.get_portfolio_transactions", _transactions)
-    portfolio_router._portfolio_service().clear_upstream_cache()
+    portfolio_service().clear_upstream_cache()
 
     client = TestClient(app)
     response = client.get(
@@ -1037,7 +1039,7 @@ def test_portfolio_insights_router_returns_blocked_exception_summaries(monkeypat
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.get_portfolio_positions", _positions)
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.query_asset_allocation", _allocation)
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.get_portfolio_transactions", _transactions)
-    portfolio_router._portfolio_service().clear_upstream_cache()
+    portfolio_service().clear_upstream_cache()
 
     client = TestClient(app)
     response = client.get("/api/v1/portfolio/portfolios/PF_1001/insights")
@@ -1175,7 +1177,7 @@ def test_portfolio_book_router(monkeypatch):
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.get_portfolio_cash_balances", _cash_balances)
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.get_cashflow_projection", _cashflow)
 
-    portfolio_router._portfolio_service().clear_upstream_cache()
+    portfolio_service().clear_upstream_cache()
     client = TestClient(app)
     response = client.get(
         "/api/v1/portfolio/portfolios/PF_1001/book",
@@ -1278,7 +1280,7 @@ def test_portfolio_transactions_router(monkeypatch):
         }
 
     monkeypatch.setattr(f"{LOTUS_CORE_QUERY_CLIENT}.get_portfolio_transactions", _transactions)
-    portfolio_router._portfolio_service().clear_upstream_cache()
+    portfolio_service().clear_upstream_cache()
     client = TestClient(app)
     response = client.get(
         "/api/v1/portfolio/portfolios/PF_1001/transactions",
@@ -1878,7 +1880,7 @@ def test_portfolio_performance_snapshot_router(monkeypatch):
         }
 
     monkeypatch.setattr(
-        portfolio_router._performance_workspace_service(),
+        portfolio_performance_workspace_service(),
         "get_portfolio_performance_snapshot",
         _snapshot,
     )
@@ -1954,7 +1956,7 @@ def test_portfolio_performance_snapshot_router_preserves_unavailable_state(monke
         }
 
     monkeypatch.setattr(
-        portfolio_router._performance_workspace_service(),
+        portfolio_performance_workspace_service(),
         "get_portfolio_performance_snapshot",
         _snapshot,
     )
